moved arraytobase64 function into a customFunctions.js in utils folder on client side, added code in classroom controller to delete temporary file from upload after saving to database
